+++ packages/fighting-test/breadcrumb.spec.ts
@@ -4,7 +4,9 @@ import { FBreadcrumb } from '@fighting-design/fighting-components'
 
 describe('FBreadcrumb', () => {
   test('class', () => {
-    const wrapper = mount(FBreadcrumb)
+    const wrapper = mount(FBreadcrumb, {
+      slots: { default: 'hello' }
+    })
     expect(wrapper.classes()).toContain('f-breadcrumb')
   })
 })