+++ vite.config.umd.ts
@@ -1,10 +1,9 @@
 import { defineConfig } from 'vite'
 import vue from '@vitejs/plugin-vue'
-import vueSetupExtend from 'vite-plugin-vue-setup-extend'
 import { resolve } from 'path'
 
 export default defineConfig({
-  plugins: [vueSetupExtend(), vue()],
+  plugins: [vue()],
   build: {
     outDir: resolve(__dirname, 'dist/dist'),
     lib: {